Add README.md
This commit adds a README.md file to the repository, providing an overview of the website, my projects, and instructions for contributing. The README.md includes sections about me, the website, projects like thumbnail design and my YouTube channel BUZnSTG, how to get in touch, and how to contribute to the project.